+++ randomized_alg/globals.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-extern int randomized_comparison_counter;
-extern int randomized_arithmetic_op_counter;
-extern int dijkstras_comparison_counter;
-extern int dijkstras_arithmetic_op_counter;
+extern long long randomized_comparison_counter;
+extern long long randomized_arithmetic_op_counter;
+extern long long dijkstras_comparison_counter;
+extern long long dijkstras_arithmetic_op_counter;
